pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "xm-playwright-pytest"
version = "1.0.0"
description = "Playwright + Python/Pytest test suite for xm.com"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.2",
    "pytest-playwright>=0.5",
    "pytest-asyncio>=0.23",
    "pytest-html>=4.1",
    "pytest-xdist>=3.5",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli       = true
log_cli_level = "INFO"
addopts       = "--timeout=90"
markers = [
    "smoke: fast pre-deployment gate",
    "regression: full regression suite",
    "websocket: WebSocket-specific tests",
    "cookie: cookie consent flow tests",
]
